+++ libraries/provider_local_win_event_log_source.rb
@@ -2,6 +2,7 @@
 
 require 'chef/provider/lwrp_base'
 require_relative 'provider_source'
+require_relative 'types'
 
 class Chef
   class Provider
@@ -11,6 +12,10 @@ class Chef
       def config_hash
         hash = super
         hash['source']['logNames'] = new_resource.log_names
+        hash['source']['eventFormat'] = EVENT_FORMAT[new_resource.event_format]
+        hash['source']['eventMessage'] = EVENT_MESSAGE[new_resource.event_message]
+        hash['source']['allowlist'] = new_resource.allowlist
+        hash['source']['denylist'] = new_resource.denylist
         hash
       end
     end