+++ libraries/provider_remote_win_event_log_source.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
 require 'chef/provider/lwrp_base'
-require_relative 'provider_source'
+require_relative 'provider_local_win_event_log_source'
 
 class Chef
   class Provider
-    class SumoSourceRemoteWindowsEventLog < Chef::Provider::SumoSource
+    class SumoSourceRemoteWindowsEventLog < Chef::Provider::SumoSourceLocalWindowsEventLog
       provides :sumo_source_remote_windows_event_log if respond_to?(:provides)
 
       def config_hash
@@ -14,7 +14,6 @@ class Chef
         hash['source']['username'] = new_resource.username
         hash['source']['password'] = new_resource.password
         hash['source']['hosts'] = new_resource.hosts
-        hash['source']['logNames'] = new_resource.log_names
         hash
       end
     end