+++ test/fixtures/cookbooks/source-resource/recipes/local_win_event_log_create.rb
@@ -3,4 +3,9 @@
 sumo_source_local_windows_event_log 'local_win_event_log' do
   source_json_directory node['sumologic']['sumo_json_path']
   log_names %w[security application]
+
+  event_format :legacy
+  event_message :message
+  allowlist "el1,el2"
+  denylist "el3,el4"
 end